Add test cases for https://github.com/ansible-collections/amazon.aws/issues/149
