+++ tests/integration/targets/s3_bucket/roles/s3_bucket/tasks/complex.yml
@@ -79,7 +79,7 @@
           - output is changed
           - output.policy.Statement[0].Action == 's3:GetObject'
           - output.policy.Statement[0].Effect == 'Deny'
-          - output.policy.Statement[0].Principal == '*'
+          - output.policy.Statement[0].Principal.AWS == '*'
           - output.policy.Statement[0].Resource == 'arn:aws:s3:::{{ bucket_name }}/*'
           - output.policy.Statement[0].Sid == 'AddPerm'
